NOREF: solve Swaption issues
