+++ scripts/configs/ff_config.json
@@ -103,7 +103,7 @@
     },
     "swaption":{
         "1":{
-            "layers": 6,
+            "layers": 12,
             "hidden_units": 64
         },
         "2":{
